backend/requirements.txt
fastapi>=0.100.0
anthropic>=0.30.0
sentence-transformers>=2.7.0
pgvector>=0.3.0
pdfplumber>=0.10.0
uvicorn>=0.22.0
gunicorn>=21.2.0
numpy>=1.24.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
ortools>=9.6.0
scipy>=1.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0
python-multipart>=0.0.6
redis>=4.5.0
shap>=0.42.0
torch>=2.0.0
prophet>=1.1.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
neo4j>=5.10.0
plotly>=5.14.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
kaleido>=0.2.1
